Fixed issue #74
Added a DestroyThySelf function in the singleton instance. pls call this function after loading the dogso in the mini game so that going back to the main menu it will have a fresh instance and won't cause any weird behavior. also added a savedogdata at the start function of getdogso so that console errors regarding saved data missing won't occur